Harden auth and stabilize local development

+++ write-service/src/main/java/com/urlshortener/write/config/KeyBlockResponse.java
@@ -6,21 +6,22 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 /**
  * Response body returned by the KGS {@code POST /internal/keys/next-block} endpoint.
  *
- * <p>The block defines an inclusive counter range {@code [startKey, endKey]}.
- * The write-service pod consumes keys from this range atomically via an
- * in-memory counter and only calls KGS again when the block is exhausted.
+ * <p>The block defines a raw counter range {@code [startKey, endKey)}.
+ * The write-service pod consumes counters from this range atomically via an
+ * in-memory counter and encodes them locally, only calling KGS again when the
+ * block is exhausted.
  */
 public record KeyBlockResponse(
         long startKey,
         long endKey,
-        int blockSize,
+        long blockSize,
         String region
 ) {
     @JsonCreator
     public KeyBlockResponse(
             @JsonProperty("startKey")  long startKey,
             @JsonProperty("endKey")    long endKey,
-            @JsonProperty("blockSize") int blockSize,
+            @JsonProperty("blockSize") long blockSize,
             @JsonProperty("region")    String region) {
         this.startKey  = startKey;
         this.endKey    = endKey;
